Add sRGB encoding gamma to output
PNG images get displayed with an sRGB colour profile by default, which
means decoding gamma is applied to the pixel values before being
displayed on screens. This causes mid tones to appear very dark. Apply
sRGB encoding gamma to the pixels written to the image to cancel out
this effect. With this change the physical brightness of the image is
linearly correlated with the number of points that hit a given area.