Add escape boundnary calculation
Add option to calculate the set of points where the z = z * z + c only
ever increases in geometric distance from the origin point.  This forms
an upper bound on the points needed to be evaluated in order to properly
draw a region of the Buddhabrot fractal.  All points part of this set
that are located further from the origin than the furthest point in the
output image can be discarded as they will never reach the image.