Track key performance metrics
Track samples and points evaluated and the time it took to do so and
print these to allow performance optimizations guided by measurements.
The key metrics to make note of is the number of starting points
evaluated (called samples) and the number of points along the paths that
escaped (called points) that were discovered.

For these numbers input is the count of elements put up for evaluation
and output is the count of elements that were kept during this
evalution. Currently for samples the ones that did not escape are
discarded and for points the ones that ended up outside the output area
are discarded.