Optimise: Stop sampling early on no output
Check that any output points have being generated from the sampling of a
mask pixel and bail early if no output is seen before a configurable
number of samples have been made.  For now this limit defaults to the
square of the total number of samples to make per mask pixel.