Remove escape boundary calculation
Remove code for calculating escape boundary as it's not needed any more.
The density plot gives a better visualisation of what points can reach a
a given bounding box on the complex plane and the insight gained from
the escape boundary, that is that points with a distance greater than
2 from the origin only increase from the origin has already been acted
upon.